refactor: 아이디 암호화
