+++ src/app/User/userDao.js
@@ -65,24 +65,3 @@ export const createNewResponse = async (requestData) => {
     return null;
   }
 };
-
-export const selectResults = async (hostId) => {
-  const user = await User.findOne({ id: hostId }).populate("friends");
-  if (user) {
-    const results = {
-      hostId: String(user.id),
-      hostName: user.name,
-      data: user.friends.map((friend) => ({
-        guestName: friend.name,
-        animal: friend.animal,
-        emoji: friend.emoji,
-        color: friend.color,
-        first: friend.first,
-        now: friend.now,
-      })),
-    };
-    return results;
-  } else {
-    return null;
-  }
-};
